Removed some duplication in tests and moved an additional extension test to extension tests
